Add token caching
Now, an access token can be cached so that
the user doesn't have to re-authenticate
everytime.

Going forward, we should maybe force the
cached token to expire after some time
period (might be that Spotify stops accepting
them after some time too?), but we'll leave
that for a future patch.